Rename warnings.ts to Warn.ts
Be consistent with the fp-ts file naming convention.
