+++ src/fp-utils.ts
@@ -3,7 +3,7 @@ import * as Either from 'fp-ts/lib/Either'
 import * as Task from 'fp-ts/lib/Task'
 import * as TaskEither from 'fp-ts/lib/TaskEither'
 
-import * as Warn from './warnings'
+import * as Warn from './Warn'
 
 export const traverseATs = Array.array.traverse(Task.taskSeq)
 export const traverseAE = Array.array.traverse(Either.either)
